Unify family dashboard with tab navigation system
- Convert individual family pages (habits, challenges, rewards, analytics, members) to tab components
- Implement unified dashboard at /dashboard/family with consistent header and theming
- Add URL parameter support for direct tab access (?tab=habits, ?tab=challenges, etc.)
- Create smart redirect system from individual pages to appropriate tabs
- Maintain all original functionality while providing seamless navigation experience
- Apply consistent blue gradient theme and dark mode support across all tabs
- Preserve touch screen optimizations and member zone functionality in overview tab
